vendor_management_project/requirements.txt
Django==4.2.6
django-ckeditor==6.7.0
django-import-export==4.0.0
django-js-asset==2.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
docx==0.2.4
drf-yasg==1.21.7